refactor: 캐러셀과 검색창을 하나의 html파일에 적용
- css는 컴포넌트별로 분리하여 작성
  : 웹컴포넌트 적용하여 온전히 재사용가능한 컴포넌트로 수정 예정
- 가독성을 위한 코드 순서 일부 조정
- html파일 이동으로 인한 경로 일부 수정
