+++ card_navigation/src/pagination.js
@@ -1,5 +1,5 @@
-import { $, delegate } from '../utils/allenibrary.js'
-import Subscriber from './subscriber.js'
+import { $, delegate } from '../../utils/allenibrary.js'
+import Subscriber from '../../utils/subscriber.js'
 
 export default class Pagination extends Subscriber {
   constructor(publisher, paginationSelector, startIdx) {
@@ -17,18 +17,6 @@ export default class Pagination extends Subscriber {
     this.delegateEvt(this.paginations.item((this.startIdx + 1) % this.paginations.length).className);
   }
 
-  delegateEvt(className) {
-    const funcMap = {
-      [className]: (target) => this.handleClick(target)
-    };
-    delegate(this.wrapper, 'click', 'className', funcMap);
-  }
-
-  handleClick(target) {
-    const targetIdx = Number(target.dataset.idx);
-    this.publisher.setState({ targetIdx });;
-  }
-
   addIdx(paginations) {
     Array.from(paginations).forEach((el, i) => {
       el.dataset.idx = i;
@@ -40,6 +28,18 @@ export default class Pagination extends Subscriber {
       if (el.classList.contains('active')) el.classList.remove('active');
       if (i === idx) el.classList.add('active');
     })
+  }
+
+  delegateEvt(className) {
+    const funcMap = {
+      [className]: (target) => this.handleClick(target)
+    };
+    delegate(this.wrapper, 'click', 'className', funcMap);
+  }
+
+  handleClick(target) {
+    const targetIdx = Number(target.dataset.idx);
+    this.publisher.setState({ targetIdx });;
   }
 
   render(state) {
