+++ search_bar/app.js
@@ -1,4 +1,3 @@
-import * as _ from '../utils/allenibrary.js'
 import SuggestionUI from './components/suggestion_ui.js';
 import RecentKeywordsUI from './components/recent_keywords_ui.js'
 import SearchBarUI from './components/search_bar_ui.js'
@@ -7,12 +6,14 @@ import { config } from './config.js'
 
 const { url, searchFormSelector, inputSelector, buttonSelector } = config;
 
-const stateManager = new StateManager({ url });
-const searchUI = new SearchBarUI(stateManager, { inputSelector, buttonSelector });
+export const initSearchBar = _ => {
+  const stateManager = new StateManager({ url });
+  const searchUI = new SearchBarUI(stateManager, { inputSelector, buttonSelector });
 
-if (config.recentKeywords) {
-  const recentKeywordsUI = new RecentKeywordsUI(stateManager, searchFormSelector);
-}
-if (config.suggestion) {
-  const suggestionUI = new SuggestionUI(stateManager, searchFormSelector);
+  if (config.recentKeywords) {
+    const recentKeywordsUI = new RecentKeywordsUI(stateManager, searchFormSelector);
+  }
+  if (config.suggestion) {
+    const suggestionUI = new SuggestionUI(stateManager, searchFormSelector);
+  }
 }
