--- search_bar/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <script src="https://kit.fontawesome.com/b7e4569d54.js"></script>
-    <link rel="stylesheet" href="style.css" />
-    <title>search bar</title>
-  </head>
-  <body>
-    <form class="search-form">
-      <select name="" id=""></select>
-      <input class="search-bar" type="text" placeholder="type here.." />
-      <button class="search-button"><i class="fas fa-search"></i></button>
-    </form>
-    <script type="module" src="app.js"></script>
-  </body>
-</html>
